Add files via upload
This is the gradient descent way of solving a linear regression problem. In sklearn LinearRegression they uses OLS method internally, that is the analytical calculation/formula based calculation hence it gives the exact solution and most accurate. Tho that is not suitable for large datasets and also for online learning, hence GD method is introduced that uses the concept of differentiation in finding the minima in the graph of the loss fuction(i.e a convex function). 
There are multiple type of GD regressor as well, this one is batch gradient regressor.